add employee util to format availability
